# backend/src/main/resources/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VieGym API",
    "description": "Backend API for VieGym workout, nutrition and AI coach",
    "contact": {
      "name": "VieGym"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development API v1"
    }
  ],
  "paths": {
...
    "/api/v1/exercises": {
      "get": {
        "tags": [
          "exercise-catalog-controller"
        ],
        "operationId": "searchExercises",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "muscleGroupId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "equipmentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "BEGINNER",
                "INTERMEDIATE",
                "ADVANCED"
              ]
            }
          },
          {
            "name": "compatibleWithMyEquipment",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name,asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseExerciseSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{id}": {
      "get": {
        "tags": [
          "exercise-catalog-controller"
        ],
        "operationId": "getExerciseDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseExerciseDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exercises/{id}/alternatives": {
      "get": {
        "tags": [
          "exercise-catalog-controller"
        ],
        "operationId": "getExerciseAlternatives",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseListExerciseSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/favorite-exercises": {
      "get": {
        "tags": [
          "exercise-catalog-controller"
        ],
        "operationId": "listFavoriteExercises",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseExerciseSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/favorite-exercises/{exerciseId}": {
      "put": {
        "tags": [
          "exercise-catalog-controller"
        ],
        "operationId": "addFavorite",
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "exercise-catalog-controller"
        ],
        "operationId": "removeFavorite",
        "parameters": [
          {
            "name": "exerciseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/personal-records": {
      "get": {
        "tags": ["personal-record-controller"],
        "operationId": "listPersonalRecords",
        "parameters": [
          { "name": "exerciseId", "in": "query", "required": false, "schema": { "type": "integer", "format": "int64" } },
          { "name": "type", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "required": false, "schema": { "type": "integer", "default": 0 } },
          { "name": "size", "in": "query", "required": false, "schema": { "type": "integer", "default": 20 } }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "*/*": { "schema": { "$ref": "#/components/schemas/ApiResponsePageResponsePersonalRecordDto" } } } }
        }
      }
    },
...
    "/api/v1/weight-logs": {
      "get": {
        "tags": [
          "weight-log-controller"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseWeightLogDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weight-logs/{loggedDate}": {
      "put": {
        "tags": [
          "weight-log-controller"
        ],
        "operationId": "upsert",
        "parameters": [
          {
            "name": "loggedDate",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertWeightLogRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUpsertWeightLogResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseUpsertWeightLogResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/weight-logs/trend": {
      "get": {
        "tags": [
          "weight-log-controller"
        ],
        "operationId": "trend",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 30
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseWeightTrendResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UpsertWeightLogRequest": {
        "type": "object",
        "properties": {
          "weightKg": {
            "type": "number",
            "maximum": 9999.99,
            "minimum": 0.01
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "weightKg"
        ]
      },
      "WeightLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "loggedDate": {
            "type": "string",
            "format": "date"
          },
          "weightKg": {
            "type": "number"
          },
          "note": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpsertWeightLogResponse": {
        "type": "object",
        "properties": {
          "weightLog": {
            "$ref": "#/components/schemas/WeightLogDto"
          },
          "metricsUpdated": {
            "type": "boolean"
          },
          "metrics": {
            "$ref": "#/components/schemas/HealthProfileResponse_Metrics"
          },
          "nutritionTargetChanged": {
            "type": "boolean"
          }
        }
      },
      "ApiResponseUpsertWeightLogResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UpsertWeightLogResponse"
          }
        }
      },
      "PageResponseWeightLogDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeightLogDto"
            }
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
      "ApiResponsePageResponseWeightLogDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseWeightLogDto"
          }
        }
      },
      "WeightTrendPoint": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "weightKg": {
            "type": "number"
          }
        }
      },
      "WeightTrendResponse": {
        "type": "object",
        "properties": {
          "days": {
            "type": "integer"
          },
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "startWeightKg": {
            "type": "number"
          },
          "currentWeightKg": {
            "type": "number"
          },
          "changeKg": {
            "type": "number"
          },
          "direction": {
            "type": "string",
            "enum": [
              "UP",
              "DOWN",
              "STABLE"
            ]
          },
          "sufficientData": {
            "type": "boolean"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeightTrendPoint"
            }
          }
        }
      },
      "ApiResponseWeightTrendResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/WeightTrendResponse"
          }
        }
      },
...
      "ExerciseMuscleGroupDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "PRIMARY",
              "SECONDARY"
            ]
          }
        }
      },
      "ExerciseEquipmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "required": {
            "type": "boolean"
          }
        }
      },
      "ExerciseSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ]
          },
          "description": {
            "type": "string"
          },
          "muscleGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseMuscleGroupDto"
            }
          },
          "equipment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseEquipmentDto"
            }
          },
          "media": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "PageResponseExerciseSummaryDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseSummaryDto"
            }
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer"
          },
          "hasNext": {
            "type": "boolean"
          }
        }
      },
      "ApiResponsePageResponseExerciseSummaryDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseExerciseSummaryDto"
          }
        }
      },
      "ApiResponseListExerciseSummaryDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseSummaryDto"
            }
          }
        }
      },
      "ExerciseDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "INTERMEDIATE",
              "ADVANCED"
            ]
          },
          "description": {
            "type": "string"
          },
          "instructionSteps": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "commonMistakes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "safetyNotes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "muscleGroups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseMuscleGroupDto"
            }
          },
          "equipment": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExerciseEquipmentDto"
            }
          },
          "media": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "ApiResponseExerciseDetailDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ExerciseDetailDto"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {}
        }
      },
...
      "PersonalRecordDto": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "exerciseId": { "type": "integer", "format": "int64" },
          "exerciseName": { "type": "string" },
          "recordType": { "type": "string" },
          "value": { "type": "number" },
          "achievedAt": { "type": "string", "format": "date-time" },
          "workoutLogId": { "type": "integer", "format": "int64" },
          "previousValue": { "type": "number" }
        }
      },
...
      "ApiResponsePageResponsePersonalRecordDto": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "object",
            "properties": {
              "content": { "type": "array", "items": { "$ref": "#/components/schemas/PersonalRecordDto" } },
              "page": { "type": "integer" },
              "size": { "type": "integer" },
              "totalElements": { "type": "integer", "format": "int64" },
              "totalPages": { "type": "integer" },
              "first": { "type": "boolean" },
              "last": { "type": "boolean" }
            }
          },
          "error": { "$ref": "#/components/schemas/ApiError" }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}